Move test_original_solution() to zebra_solver.py
