+++ src/zebra_puzzles/clue_selection.py
@@ -9,7 +9,7 @@ from zebra_puzzles.clue_removal import (
     remove_redundant_clues_with_rules,
     remove_redundant_clues_with_solver,
 )
-from zebra_puzzles.zebra_solver import format_solution_as_matrix, solver
+from zebra_puzzles.zebra_solver import solver, test_original_solution
 from zebra_puzzles.zebra_utils import describe_random_attributes
 
 
@@ -146,35 +146,6 @@ def choose_clues(
             raise StopIteration("Used too many attempts to solve the puzzle.")
 
     return chosen_clues, chosen_clue_types
-
-
-def test_original_solution(
-    solutions: list[dict[str, int]],
-    solution: np.ndarray,
-    n_objects: int,
-    n_attributes: int,
-    chosen_clues: list[str],
-):
-    """Test if the solver found the original solution or an unexpected one.
-
-    Finding a new solution should not be possible and indicates a bug in the solver or the clue selection process. If this happens, an error is raised.
-
-    Args:
-        solutions: Solutions to the zebra puzzle found by the solver as a list of dictionaries containing object indices and chosen attribute values.
-        solution: Expected solution to the zebra puzzle as a matrix of strings containing object indices and chosen attribute values. This matrix is n_objects x (n_attributes + 1).
-        n_objects: Number of objects in the puzzle as an integer.
-        n_attributes: Number of attributes per object as an integer.
-        chosen_clues: Clues for the zebra puzzle as a list of strings.
-
-    """
-    solution_attempt = format_solution_as_matrix(
-        solution_dict=solutions[0], n_objects=n_objects, n_attributes=n_attributes
-    )
-
-    if [sorted(obj) for obj in solution_attempt] != [sorted(obj) for obj in solution]:
-        raise ValueError(
-            f"The solver has found a solution that is not the expected one: \nFound \n{solution_attempt} \nExpected \n{solution} \nChosen clues: \n{chosen_clues}"
-        )
 
 
 def get_clue_probabilities(
